+++ src/ska_mid_cbf_mcs/power_switch/power_switch_device.py
@@ -109,6 +109,7 @@ class PowerSwitch(CbfDevice):
             logger=self.logger,
             communication_state_callback=self._communication_state_changed,
             component_state_callback=self._component_state_changed,
+            admin_mode_callback=self._admin_mode_perform_action,
         )
 
     def init_command_objects(self: PowerSwitch) -> None: