+++ src/ska_mid_cbf_mcs/talon_lru/talon_lru_device.py
@@ -77,6 +77,7 @@ class TalonLRU(CbfDevice):
             health_state_callback=self._update_health_state,
             communication_state_callback=self._communication_state_changed,
             component_state_callback=self._component_state_changed,
+            admin_mode_callback=self._admin_mode_perform_action,
         )
 
     # -------------