+++ tests/test_utils.py
@@ -22,7 +22,7 @@ def device_online_and_on(
     :param event_tracer: A :py:class:`TangoEventTracer` used to recieve subscribed change events from the device under test.
     """
     # Set a given device to AdminMode.ONLINE and DevState.ON
-    device_under_test.simulationMode == SimulationMode.FALSE
+    device_under_test.simulationMode = SimulationMode.FALSE
     device_under_test.adminMode = AdminMode.ONLINE
 
     assert_that(event_tracer).within_timeout(